src/genie-corewebhook-cf/requirements.txt
functions-framework==3.*
flask
google-cloud-discoveryengine=0.13.6